requirements.txt
google-generativeai
customtkinter
rembg
onnxruntime
Pillow
opencv-python
python-dotenv
